// client/stocklinx app/src/redux/asset/constant.ts
export enum assetConst {
  FETCH_ASSETS_REQUEST = "FETCH_ASSETS_REQUEST",
  FETCH_ASSETS_SUCCESS = "FETCH_ASSETS_SUCCESS",
  FETCH_ASSETS_FAILURE = "FETCH_ASSETS_FAILURE",
  FETCH_ASSET_REQUEST = "FETCH_ASSET_REQUEST",
  FETCH_ASSET_SUCCESS = "FETCH_ASSET_SUCCESS",
  FETCH_ASSET_FAILURE = "FETCH_ASSET_FAILURE",
  FETCH_ASSET_COUNTS_REQUEST = "FETCH_ASSET_COUNTS_REQUEST",
  FETCH_ASSET_COUNTS_SUCCESS = "FETCH_ASSET_COUNTS_SUCCESS",
  FETCH_ASSET_COUNTS_FAILURE = "FETCH_ASSET_COUNTS_FAILURE",
  CREATE_ASSET_REQUEST = "CREATE_ASSET_REQUEST",
  CREATE_ASSET_SUCCESS = "CREATE_ASSET_SUCCESS",
  CREATE_ASSET_FAILURE = "CREATE_ASSET_FAILURE",
  CREATE_RANGE_ASSET_REQUEST = "CREATE_RANGE_ASSET_REQUEST",
  CREATE_RANGE_ASSET_SUCCESS = "CREATE_RANGE_ASSET_SUCCESS",
  CREATE_RANGE_ASSET_FAILURE = "CREATE_RANGE_ASSET_FAILURE",
  UPDATE_ASSET_REQUEST = "UPDATE_ASSET_REQUEST",
  UPDATE_ASSET_SUCCESS = "UPDATE_ASSET_SUCCESS",
  UPDATE_ASSET_FAILURE = "UPDATE_ASSET_FAILURE",
  REMOVE_ASSET_REQUEST = "REMOVE_ASSET_REQUEST",
  REMOVE_ASSET_SUCCESS = "REMOVE_ASSET_SUCCESS",
  REMOVE_ASSET_FAILURE = "REMOVE_ASSET_FAILURE",
  REMOVE_RANGE_ASSET_REQUEST = "REMOVE_RANGE_ASSET_REQUEST",
  REMOVE_RANGE_ASSET_SUCCESS = "REMOVE_RANGE_ASSET_SUCCESS",
  REMOVE_RANGE_ASSET_FAILURE = "REMOVE_RANGE_ASSET_FAILURE",
  SET_ASSET = "SET_ASSET",
  CLEAR_ASSET = "CLEAR_ASSET",
  SET_ASSETS = "SET_ASSETS",
  CLEAR_ASSETS = "CLEAR_ASSETS",
}